chore: rename evalctx to gframe
The name "evaluation context" for frames are very bad. This name comes from a
misunderstanding of mine when writing the original patch.

The thing should really be called a "frame", but we already have "frame"s for
function call frames. Sam proposes "gframe" for "general frame", but there
might be better names for it.

(Previously I also called them "context"; that was also misleading.)