refactor(al): add RestoreI
RestoreI is somewhat similar to ExecuteSeqI in the sense that they both
execute a sequence of wasm instruction in a point in the future.
Their difference is that ExecuteSeqI executes the instructions immediately,
pausing AL execution before the wasm instructions finishes, while RestoreI
simply pushes the instructions to the wasm instruction stack, and continues
AL execution immediately.

RestoreI is not used anywhere now; it is for stack-switching.